UPSTREAM: 1468: Fixed parsing of /proc/self/mountinfo with spaces
When a mount point contains space, it is escaped in /proc/self/mountifo as
\040:

7331 6783 8:32 / /var/lib/kubelet/plugins/kubernetes.io/csi/pv/csi.vsphere.vmware.com-[WorkloadDatastore]\0405137595f-7ce3-e95a-5c03-06d835dea807/e2e-vmdk-1641374604660540311.vmdk/globalmount rw,relatime - ext4 /dev/sdc rw,seclabel

Un-escape this sequence (and any other \nnn sequences) before comparing the
parsed mount path.

This happens only for in-line in-tree vSphere volumes when CSI migration is
enabled. Whole '[WorkloadDatastore] 5137595f-7ce3-e95a-5c03-06d835dea807/e2e-vmdk-1641374604660540311.vmdk'
is used as volumeHandle (and thus in global mount dir) and a fake PV name (and
thus in pod local mount dir).